split code highlighting plugin
