+++ packages/lexical-code/src/CodeHighlighter.ts
@@ -6,19 +6,17 @@
  *
  */
 
+// eslint-disable-next-line simple-import-sort/imports
 import type {
   DOMConversionMap,
   DOMConversionOutput,
   EditorConfig,
-  EditorThemeClasses,
-  LexicalCommand,
   LexicalEditor,
   LexicalNode,
   NodeKey,
   ParagraphNode,
   RangeSelection,
   SerializedElementNode,
-  SerializedTextNode,
 } from 'lexical';
 
 import * as Prism from 'prismjs';
@@ -35,14 +33,7 @@ import 'prismjs/components/prism-python';
 import 'prismjs/components/prism-rust';
 import 'prismjs/components/prism-swift';
 
-import {registerCodeIndent} from "./EditorShortcuts";
-
-import {
-  addClassNamesToElement,
-  mergeRegister,
-  removeClassNamesFromElement,
-} from '@lexical/utils';
-
+import {addClassNamesToElement, mergeRegister} from '@lexical/utils';
 import {
   $createLineBreakNode,
   $createParagraphNode,
@@ -52,17 +43,17 @@ import {
   $isLineBreakNode,
   $isRangeSelection,
   $isTextNode,
-  COMMAND_PRIORITY_LOW,
   ElementNode,
-  INDENT_CONTENT_COMMAND,
-  KEY_ARROW_DOWN_COMMAND,
-  KEY_ARROW_UP_COMMAND,
-  MOVE_TO_START,
-  MOVE_TO_END,
-  OUTDENT_CONTENT_COMMAND,
   TextNode,
 } from 'lexical';
 import {Spread} from 'libdefs/globals';
+
+import {registerCodeIndent} from './EditorShortcuts';
+import {
+  $createCodeHighlightNode,
+  $isCodeHighlightNode,
+  CodeHighlightNode,
+} from './HighlighterHelper';
 
 const DEFAULT_CODE_LANGUAGE = 'javascript';
 
@@ -75,17 +66,6 @@ type SerializedCodeNode = Spread<
   SerializedElementNode
 >;
 
-type SerializedCodeHighlightNode = Spread<
-  {
-    highlightType: string | null | undefined;
-    type: 'code-highlight';
-    version: 1;
-  },
-  SerializedTextNode
->;
-
-
-
 const mapToPrismLanguage = (
   language: string | null | undefined,
 ): string | null | undefined => {
@@ -94,97 +74,6 @@ const mapToPrismLanguage = (
     ? language
     : undefined;
 };
-
-export class CodeHighlightNode extends TextNode {
-  __highlightType: string | null | undefined;
-
-  constructor(text: string, highlightType?: string, key?: NodeKey) {
-    super(text, key);
-    this.__highlightType = highlightType;
-  }
-
-  static getType(): string {
-    return 'code-highlight';
-  }
-
-  static clone(node: CodeHighlightNode): CodeHighlightNode {
-    return new CodeHighlightNode(
-      node.__text,
-      node.__highlightType || undefined,
-      node.__key,
-    );
-  }
-
-  getHighlightType(): string | null | undefined {
-    const self = this.getLatest();
-    return self.__highlightType;
-  }
-
-  createDOM(config: EditorConfig): HTMLElement {
-    const element = super.createDOM(config);
-    const className = getHighlightThemeClass(
-      config.theme,
-      this.__highlightType,
-    );
-    addClassNamesToElement(element, className);
-    return element;
-  }
-
-  updateDOM(
-    prevNode: CodeHighlightNode,
-    dom: HTMLElement,
-    config: EditorConfig,
-  ): boolean {
-    const update = super.updateDOM(prevNode, dom, config);
-    const prevClassName = getHighlightThemeClass(
-      config.theme,
-      prevNode.__highlightType,
-    );
-    const nextClassName = getHighlightThemeClass(
-      config.theme,
-      this.__highlightType,
-    );
-    if (prevClassName !== nextClassName) {
-      if (prevClassName) {
-        removeClassNamesFromElement(dom, prevClassName);
-      }
-      if (nextClassName) {
-        addClassNamesToElement(dom, nextClassName);
-      }
-    }
-    return update;
-  }
-
-  static importJSON(
-    serializedNode: SerializedCodeHighlightNode,
-  ): CodeHighlightNode {
-    const node = $createCodeHighlightNode(serializedNode.highlightType);
-    node.setFormat(serializedNode.format);
-    node.setDetail(serializedNode.detail);
-    node.setMode(serializedNode.mode);
-    node.setStyle(serializedNode.style);
-    return node;
-  }
-
-  exportJSON(): SerializedCodeHighlightNode {
-    return {
-      ...super.exportJSON(),
-      highlightType: this.getHighlightType(),
-      type: 'code-highlight',
-    };
-  }
-
-  // Prevent formatting (bold, underline, etc)
-  setFormat(format: number): this {
-    return this;
-  }
-}
-
-export function $isCodeHighlightNode(
-  node: LexicalNode | CodeHighlightNode | null | undefined,
-): node is CodeHighlightNode {
-  return node instanceof CodeHighlightNode;
-}
 
 export function getFirstCodeHighlightNodeOfLine(
   anchor: LexicalNode,
@@ -224,24 +113,63 @@ export function getLastCodeHighlightNodeOfLine(
   return currentNode;
 }
 
-export function $createCodeHighlightNode(
-  text: string,
-  highlightType?: string,
-): CodeHighlightNode {
-  return new CodeHighlightNode(text, highlightType);
+function convertPreElement(domNode: Node): DOMConversionOutput {
+  return {node: $createCodeNode()};
 }
 
-function getHighlightThemeClass(
-  theme: EditorThemeClasses,
-  highlightType: string | undefined,
-): string | undefined {
-  return (
-    highlightType &&
-    theme &&
-    theme.codeHighlight &&
-    theme.codeHighlight[highlightType]
-  );
+function convertDivElement(domNode: Node): DOMConversionOutput {
+  // domNode is a <div> since we matched it by nodeName
+  const div = domNode as HTMLDivElement;
+  return {
+    after: (childLexicalNodes) => {
+      const domParent = domNode.parentNode;
+      if (domParent != null && domNode !== domParent.lastChild) {
+        childLexicalNodes.push($createLineBreakNode());
+      }
+      return childLexicalNodes;
+    },
+    node: isCodeElement(div) ? $createCodeNode() : null,
+  };
 }
+
+function convertTableElement(): DOMConversionOutput {
+  return {node: $createCodeNode()};
+}
+
+function convertCodeNoop(): DOMConversionOutput {
+  return {node: null};
+}
+
+function convertTableCellElement(domNode: Node): DOMConversionOutput {
+  // domNode is a <td> since we matched it by nodeName
+  const cell = domNode as HTMLTableCellElement;
+
+  return {
+    after: (childLexicalNodes) => {
+      if (cell.parentNode && cell.parentNode.nextSibling) {
+        // Append newline between code lines
+        childLexicalNodes.push($createLineBreakNode());
+      }
+      return childLexicalNodes;
+    },
+    node: null,
+  };
+}
+
+function isCodeElement(div: HTMLDivElement): boolean {
+  return div.style.fontFamily.match('monospace') !== null;
+}
+
+function isGitHubCodeCell(
+  cell: HTMLTableCellElement,
+): cell is HTMLTableCellElement {
+  return cell.classList.contains('js-file-line');
+}
+
+function isGitHubCodeTable(table: HTMLTableElement): table is HTMLTableElement {
+  return table.classList.contains('js-file-line-container');
+}
+
 const LANGUAGE_DATA_ATTRIBUTE = 'data-highlight-language';
 
 export class CodeNode extends ElementNode {
@@ -644,7 +572,11 @@ function getDiffRange(
 // in both cases we'll rerun whole reformatting over CodeNode, which is redundant.
 // Especially when pasting code into CodeBlock.
 let isHighlighting = false;
-function codeNodeTransform(node: CodeNode, editor: LexicalEditor, threshold?: number) {
+function codeNodeTransform(
+  node: CodeNode,
+  editor: LexicalEditor,
+  threshold?: number,
+) {
   if (isHighlighting) {
     return;
   }
@@ -669,7 +601,10 @@ function codeNodeTransform(node: CodeNode, editor: LexicalEditor, threshold?: nu
         const highlightNodes = getHighlightNodes(tokens);
         const diffRange = getDiffRange(node.getChildren(), highlightNodes);
         const {from, to, nodesForReplacement} = diffRange;
-        if ((from !== to || nodesForReplacement.length) && code.length <= threshold) {
+        if (
+          (from !== to || nodesForReplacement.length) &&
+          code.length <= threshold
+        ) {
           node.splice(from, to - from, nodesForReplacement);
           return true;
         }
@@ -685,7 +620,11 @@ function codeNodeTransform(node: CodeNode, editor: LexicalEditor, threshold?: nu
   );
 }
 
-function textNodeTransform(node: TextNode, editor: LexicalEditor, threshold?: number): void {
+function textNodeTransform(
+  node: TextNode,
+  editor: LexicalEditor,
+  threshold?: number,
+): void {
   // Since CodeNode has flat children structure we only need to check
   // if node's parent is a code node and run highlighting if so
   const parentNode = node.getParent();
@@ -698,212 +637,10 @@ function textNodeTransform(node: TextNode, editor: LexicalEditor, threshold?: nu
   }
 }
 
-function doIndent(node: CodeHighlightNode, type: LexicalCommand<void>) {
-  const text = node.getTextContent();
-  if (type === INDENT_CONTENT_COMMAND) {
-    // If the codeblock node doesn't start with whitespace, we don't want to
-    // naively prepend a '\t'; Prism will then mangle all of our nodes when
-    // it separates the whitespace from the first non-whitespace node. This
-    // will lead to selection bugs when indenting lines that previously
-    // didn't start with a whitespace character
-    if (text.length > 0 && /\s/.test(text[0])) {
-      node.setTextContent('\t' + text);
-    } else {
-      const indentNode = $createCodeHighlightNode('\t');
-      node.insertBefore(indentNode);
-    }
-  } else {
-    if (text.indexOf('\t') === 0) {
-      // Same as above - if we leave empty text nodes lying around, the resulting
-      // selection will be mangled
-      if (text.length === 1) {
-        node.remove();
-      } else {
-        node.setTextContent(text.substring(1));
-      }
-    }
-  }
-}
-
-function handleMultilineIndent(type: LexicalCommand<void>): boolean {
-  const selection = $getSelection();
-
-  if (!$isRangeSelection(selection) || selection.isCollapsed()) {
-    return false;
-  }
-
-  // Only run multiline indent logic on selections exclusively composed of code highlights and linebreaks
-  const nodes = selection.getNodes();
-  for (let i = 0; i < nodes.length; i++) {
-    const node = nodes[i];
-    if (!$isCodeHighlightNode(node) && !$isLineBreakNode(node)) {
-      return false;
-    }
-  }
-  const startOfLine = getFirstCodeHighlightNodeOfLine(nodes[0]);
-
-  if (startOfLine != null) {
-    doIndent(startOfLine, type);
-  }
-
-  for (let i = 1; i < nodes.length; i++) {
-    const node = nodes[i];
-    if ($isLineBreakNode(nodes[i - 1]) && $isCodeHighlightNode(node)) {
-      doIndent(node, type);
-    }
-  }
-
-  return true;
-}
-
-function handleShiftLines(
-  type: LexicalCommand<KeyboardEvent>,
-  event: KeyboardEvent,
-): boolean {
-  // We only care about the alt+arrow keys
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) {
-    return false;
-  }
-
-  // I'm not quite sure why, but it seems like calling anchor.getNode() collapses the selection here
-  // So first, get the anchor and the focus, then get their nodes
-  const {anchor, focus} = selection;
-  const anchorOffset = anchor.offset;
-  const focusOffset = focus.offset;
-  const anchorNode = anchor.getNode();
-  const focusNode = focus.getNode();
-  const arrowIsUp = type === KEY_ARROW_UP_COMMAND;
-
-  // Ensure the selection is within the codeblock
-  if (!$isCodeHighlightNode(anchorNode) || !$isCodeHighlightNode(focusNode)) {
-    return false;
-  }
-  if (!event.altKey) {
-    // Handle moving selection out of the code block, given there are no
-    // sibling thats can natively take the selection.
-    if (selection.isCollapsed()) {
-      const codeNode = anchorNode.getParentOrThrow();
-      if (
-        arrowIsUp &&
-        anchorOffset === 0 &&
-        anchorNode.getPreviousSibling() === null
-      ) {
-        const codeNodeSibling = codeNode.getPreviousSibling();
-        if (codeNodeSibling === null) {
-          codeNode.selectPrevious();
-          event.preventDefault();
-          return true;
-        }
-      } else if (
-        !arrowIsUp &&
-        anchorOffset === anchorNode.getTextContentSize() &&
-        anchorNode.getNextSibling() === null
-      ) {
-        const codeNodeSibling = codeNode.getNextSibling();
-        if (codeNodeSibling === null) {
-          codeNode.selectNext();
-          event.preventDefault();
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  const start = getFirstCodeHighlightNodeOfLine(anchorNode);
-  const end = getLastCodeHighlightNodeOfLine(focusNode);
-  if (start == null || end == null) {
-    return false;
-  }
-
-  const range = start.getNodesBetween(end);
-  for (let i = 0; i < range.length; i++) {
-    const node = range[i];
-    if (!$isCodeHighlightNode(node) && !$isLineBreakNode(node)) {
-      return false;
-    }
-  }
-
-  // After this point, we know the selection is within the codeblock. We may not be able to
-  // actually move the lines around, but we want to return true either way to prevent
-  // the event's default behavior
-  event.preventDefault();
-  event.stopPropagation(); // required to stop cursor movement under Firefox
-
-  const linebreak = arrowIsUp
-    ? start.getPreviousSibling()
-    : end.getNextSibling();
-  if (!$isLineBreakNode(linebreak)) {
-    return true;
-  }
-  const sibling = arrowIsUp
-    ? linebreak.getPreviousSibling()
-    : linebreak.getNextSibling();
-  if (sibling == null) {
-    return true;
-  }
-
-  const maybeInsertionPoint = arrowIsUp
-    ? getFirstCodeHighlightNodeOfLine(sibling)
-    : getLastCodeHighlightNodeOfLine(sibling);
-  let insertionPoint =
-    maybeInsertionPoint != null ? maybeInsertionPoint : sibling;
-  linebreak.remove();
-  range.forEach((node) => node.remove());
-  if (type === KEY_ARROW_UP_COMMAND) {
-    range.forEach((node) => insertionPoint.insertBefore(node));
-    insertionPoint.insertBefore(linebreak);
-  } else {
-    insertionPoint.insertAfter(linebreak);
-    insertionPoint = linebreak;
-    range.forEach((node) => {
-      insertionPoint.insertAfter(node);
-      insertionPoint = node;
-    });
-  }
-
-  selection.setTextNodeRange(anchorNode, anchorOffset, focusNode, focusOffset);
-
-  return true;
-}
-
-function handleMoveTo(
-  type: LexicalCommand<KeyboardEvent>,
-  event: KeyboardEvent,
-): boolean {
-  const selection = $getSelection();
-  if (!$isRangeSelection(selection)) {
-    return false;
-  }
-
-  const {anchor, focus} = selection;
-  const anchorNode = anchor.getNode();
-  const focusNode = focus.getNode();
-  const isMoveToStart = type === MOVE_TO_START;
-
-  if (!$isCodeHighlightNode(anchorNode) || !$isCodeHighlightNode(focusNode)) {
-    return false;
-  }
-
-  let node;
-  let offset;
-
-  if (isMoveToStart) {
-    ({node, offset} = getStartOfCodeInLine(focusNode));
-  } else {
-    ({node, offset} = getEndOfCodeInLine(focusNode));
-  }
-
-  if (node !== null && offset !== -1) {
-    selection.setTextNodeRange(node, offset, node, offset);
-  }
-
-  event.preventDefault();
-  event.stopPropagation();
-}
-
-export function registerCodeHighlighting(editor: LexicalEditor, threshold: number): () => void {
+export function registerCodeHighlighting(
+  editor: LexicalEditor,
+  threshold: number,
+): () => void {
   if (!editor.hasNodes([CodeNode, CodeHighlightNode])) {
     throw new Error(
       'CodeHighlightPlugin: CodeNode or CodeHighlightNode not registered on editor',
